feat: move repository to db/repositories
